+++ monoloco/network/process.py
@@ -78,9 +78,9 @@ def factory_for_gt(im_size, focal_length=5.7, name=None, path_gt=None):
 
     # Without ground-truth-file
     elif im_size[0] / im_size[1] > 2.5:  # KITTI default
-            kk = [[718.3351, 0., 600.3891], [0., 718.3351, 181.5122], [0., 0., 1.]]  # Kitti calibration
-            dic_gt = None
-            logger.info("Using KITTI calibration matrix...")
+        kk = [[718.3351, 0., 600.3891], [0., 718.3351, 181.5122], [0., 0., 1.]]  # Kitti calibration
+        dic_gt = None
+        logger.info("Using KITTI calibration matrix...")
     else:  # nuScenes camera parameters
         kk = [
             [im_size[0]*focal_length/Sx, 0., im_size[0]/2],
